+++ Kafka_Example_01_API/Core/DTOs/UpdateQuantityDTO.cs
@@ -2,8 +2,8 @@
 {
     public class UpdateQuantityDTO
     {
-        public int  ProductId { get; set; }
-        public int  Quantity  { get; set; }
-        public bool Increase  { get; set; }
+        public decimal  ProductId { get; set; }
+        public decimal  Quantity  { get; set; }
+        public bool     Increase  { get; set; }
     }
 }
